requirements.txt
openai
python-dotenv
pyaudio
sounddevice
soxr
scipy
numpy
textual
pvporcupine
openai[realtime]